Fix wxPython 4.x text rendering and UI sizing issues
- Fix DrawText coordinate type errors: wxPython 4.x requires integer coordinates, not floats
  - vlist.py: Fix action list label and summary rendering
  - imageInspector.py: Fix image inspector label rendering
  - toasterbox.py: Fix notification popup text coordinates
- Increase default window width from 400 to 600 pixels to display all toolbar icons
- Add explicit font setting to tree control for better text rendering

These fixes resolve the issue where action labels were not displaying in the main window.

+++ phatch/other/pyWx/toasterbox.py
@@ -793,11 +793,11 @@ class ToasterBoxWindow(wx.Frame):
        for line in lines:
            w, h = dc.GetTextExtent(line)
            fh += h + textPadding
-       y = (th - fh) / 2; coords = []
+       y = int((th - fh) / 2); coords = []
 
        for line in lines:
            w, h = dc.GetTextExtent(line)
-           x = (tw - w) / 2
+           x = int((tw - w) / 2)
            coords.append((x, y))
            y += h + textPadding
